refactor: categoryPage 폴더 변경
