+++ src/components/routes/category/RegisterByCategory/InformationGroup.tsx
@@ -1,6 +1,6 @@
 import { RegisterInfoByCategory } from '@/model/information';
-import MonsterCardList from '../../common/MonsterCardList/MonsterCardList';
-import Registration from '../../common/Registration/Registration';
+import MonsterCardList from '../../../common/MonsterCardList/MonsterCardList';
+import Registration from '../../../common/Registration/Registration';
 import styles from './css/InformationGroup.module.css';
 
 type Props = {
