+++ src/components/routes/category/RegisterByCategory/css/InformationGroup.module.css
